[test] Fix test for cases where TMPDIR shares a prefix with CWD (#74741)
There is code in `IRGenDebugInfo.cpp` ([1]) that tries to move the
common part of the `filename` into the `directory`. If this test is
build with a temporary directory that shares a prefix with the current
working directory, the common part of the path will be moved into
`directory`, but since the test checks for an empty directory, it will
fail the test, even if the results are correct. This situation is not
uncommon in CI setups where each job might have an isolated temporary
directory sharing a prefix with the current working directory (example:
`/ci/build/XXX/tmp` and `/ci/build/XXX/cwd`).

Modify the regex to allow anything in the `directory` value, since it is
not really important for the test to prove that the debug information is
generated.`

[1]: https://github.com/swiftlang/swift/blob/9c4232c89d8adf9068efb53105e95d14a00b2505/lib/IRGen/IRGenDebugInfo.cpp?fbclid=IwZXh0bgNhZW0CMTEAAR1eg4aiepWRUGYPzTe4KnjDId-PmyM0CYsG6wHOKTgXwvPKDP46BFWOU4U_aem_1A9mSFYy41hgN1V6RijBKA#L564-L585