[Concurrency] Make deinit check test not racy
The test is checking that the task executor deinit runs after task
completes. This deinit is caused on a different thread, and thus is
racing with the main() completing. This is racy because as main() exits,
we just exit the process and may not see the deinit get printed, making
the test look like the deinit "didnt trigger" but it did, just was
"late to the party"

resolves rdar://130528198