[cxx-interop] noexcept specifier before function attributes
In #74516 the `SWIFT_NOEXCEPT` specifier is added to the imported Swift
functions in C++ mode, but it is added after the function attributes. It seems
that the tests only do `-fsyntax-only`, which seems not to catch an error like
"expected function body after function declarator" when the header is
used without that flag.

Flip the attributes and the specifier around in the printer, and flip
them in all the tests.

(cherry picked from commit 1ea7c5a865dbbd794e243c7b0e88d4a7a049df9b)