[Sema] Fix an IsPattern optional unwrap logic error
The case being handled is the input being more
optional than the cast type, so we need to inject
optional patterns to bring the input down to the
same optionality of the cast. Therefore we need
to take the optional types from the start of the
input optional types, not the end.

Luckily this issue doesn't currently cause any
issues because `coercePatternToType` will fixup
the right type.