[ClangImporter] Adjust tests after Clang changes for `__attribute__((pure))`
There was a recent change in Clang's handling of `__attribute__((pure))`: https://github.com/llvm/llvm-project/pull/78200

`__attribute__((pure))` is now dropped for const functions, and for functions that return `void`.

This adjusts the test for the new behavior. There might be future changes required in the Swift's handling of `__attribute__((pure))`.

rdar://131300381