[Concurrency] TaskExecutors may be non-swift objects; dont swift_release them
Since we introduced proper ownership of task executors, they are now
released and retained. The problem appears with a dispatch_queue_t which
conforms to TaskExecutor being passed to Task initializer and retains
work okey because it is __owned. However, upon destroy we swift_released
the executor reference, which is incorrect as we must be using object
specific release methods -- in this case the safe way to support all
kinds of objects is `swift_unknownObjectRelease`

resolves rdar://131151645