[Concurrency] Don't infer actor isolation from inherited conformances.
If a conformance is inherited from a superclass, the isolation
of the subclass should be inferred directly from the superclass.
If the superclass has opted out of global actor inference from a
protocol, such as by conforming to the protocol in an extension,
then the subclass should not infer isolation from the protocol.

(cherry picked from commit 86f3fe1680e6bb7078afc9dbc6739ed5696c3e49)