[Casting] Structs cannot conform to NSObjectProtocol
Casting a struct to NSObjectProtocol should always fail.
This was partly fixed last year, but there was still a hole
that would successfully cast an Optional<S> to NSObjectProtocol
because the cast logic in this case lost track of the destination
type.

More specifically, `tryCastFromObjCBridgeableToClass` was
also being used to cast to class existentials.  These
two cases (casting to a class or casting to a class-constrained
existential) both involve the same bridge logic, but then
need to follow it with an additional step to get to the
final type.  Creating a new `tryCastFromObjCBridgeableToClassExistential`
made it easy to separate these cases.

Resolves rdar://107559117