[Interface] Always build swiftinterface when explicitly requested
When explicitly requested to build a swiftinterface file to swiftmodule,
alway perform the entire compilation, instead of immitating implicit
build to do a up-to-date check in order to skip the compilation.

This can cause different build behavior depending on the order of the
task, turning a failed compilation into a successful one. The next step
might be using a stricter module hash to avoid output path collision all
together.

rdar://131780720