Downgrade the error about `#isolation` in `@_unsafeInheritExecutor` when the concurrency library is old
The introduction of the error concerning use of `#isolation` within an
`@_unsafeInheritExecutor` function was carefully staged between the
concurrency library and compiler to avoid breaking source
compatibility. However, it is a source compatibility break when
dealing certain revisions of the concurrency library that adopted
`#isolation` prior to the compiler error. Detect when the _Concurrency
library doesn't have the necessary workarounds in it and downgrade the
error to a warning.

Fixes rdar://132023573.